customer_support_automation/requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
crewai-tools>=0.1.0
langchain_community==0.0.29
openai>=0.27.0
requests>=2.28.0
python-dotenv>=0.21.0
rich>=12.0
Markdown>=3.4.1
selenium>=4.8.0
webdriver-manager>=3.8.5
fastapi
uvicorn
streamlit
requests
setuptools>=65.0.0